add tests for synthese + separate classes in different files + separate tests per module + kwargs can now use CLater
