+++ backend/geonature/tests/benchmarks/__init__.py
@@ -1,0 +1,3 @@
+# Import required for CLater class when using eval()
+from flask import url_for
+from .benchmark_data import *
